US2 - refactoring - Take off usecase is more explicit in flow - removePlane and addPlane from airport inventory service return void, as this is a command operation - for planeInventory() use a copy and make it unmodifiable so it keeps the list immutable - better naming

+++ src/main/java/com/hanfak/airport/usecase/TakeOffUseCase.java
@@ -12,6 +12,10 @@ public class TakeOffUseCase {
 
   // What to return??? type - id, status, inHanger, inAirport,
   public boolean instructPlaneToTakeOff(Plane plane) {
-    return planeInventoryService.removePlane(plane);
+    if (planeInventoryService.checkPlaneIsAtAirport(plane.planeId)) {
+      planeInventoryService.removePlane(plane);
+      return true; // replace booleans with call to planeInventoryService.checkPlaneIsAtAirport(plane.planeId)
+    }
+    return false;
   }
 }
